+++ dev/run_all_test.R
@@ -6,7 +6,7 @@ rm(list = ls())
 devtools::load_all('.')
 devtools::document('.')
 # attachment::att_amend_desc()
-devtools::test()
+devtools::test(stop_on_failure = FALSE)
 devtools::check()
 
 # This skip all tests...